Put explanatory messages into our "empty" target files
As an aid for people not familiar with empty target files, we now add
an informative message to each such file, explaining what it is used
for, and where to go for additional background.

Now, each file contains a message like this:

```
This is an "empty target" file; see
https://www.gnu.org/software/make/manual/html_node/Empty-Targets.html.

It is generated by the 'build-image' target in

        /home/maier/src/grapl-security/grapl/src/js/engagement_view/Makefile

Please consult that target that for additional context.
```

The name of the target and the path to the Makefile in question are
modified as needed to suit each individual target.

(While empty target files are canonically empty, their function does
not _require_ them to be empty. The only thing Make cares about is the
modification timestamp of the file.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>