Use host's private IP as dns server for containers
With the move of dnsmasq to the host, we now need a way to tell the
containers to use the host for their dns lookups if they want to use
consul dns. Typically on a host you would just add a nameserver for
127.0.0.1, but that doesn't work because in the container 127.0.0.1
points to the container not the host.
Ideally we'd just be passing in attr.unique.network.ip-address,
but unfortunately, variable interpolation is only available for network.dns
as of Nomad 1.3 per https://github.com/hashicorp/nomad/issues/11851
So instead we're stuck doing grabbing an IP from `hostname -I` and
passing that in as an environment variable. This may actually be
for the best if we end up setting a distinct dns service in production