Download the tc-redirect-tap plugin from cloudsmith
Instead of compiling this from source which requires golang installed
and set up properly, we're just throwing this into cloudsmith for the
moment and downloading it directly from there. Long-term, either
the project will create github releases or we can fork and create the
releases ourselves. However, this changes rarely (4 merged PRs in all
of 2021), so we can update this manually ourselves if necessary