Set up systemd-resolve for dns
This has the start for a switch to systemd-resolve. The plus side is
that systemd-resolve comes pre-installed on Debian Bullseye. The down
side is that it has its own set of Nomad-related bugs and basically
needs a similar fix as dnsmasq due to docker stripping out loopback
addresses. And of course since its on 127.0.0.53, we need to explicitly
 set another dns lister so that containers can reach it from inside
bridge networks.