pipeline-ingress service (#1579)
Add pipeline-ingress API

This PR simplifies the pipeline ingress API stubs because the API does not
actually need to be bidirectionally streaming. For the current testing purposes
it can be synchronous.

The pipeline-ingress API is implemented in two layers. The underlying protocol
layer is encapsulated in rust-proto-new. No generated protocol buffer stubs are
exposed. The business logic lives in pipeline-ingress.

The protocol layer is exercised by unit tests. A new testing pipeline
test-integration-new exercises the pipeline-ingress integration tests in such a
way that they can work both locally and against AWS.