Pass in dns nameserver
Contrary to the dnsmasq logs it does not bind to the nameserver ip,
just the private ip (which makes sense, since otherwise it would need
to update extra_hosts for everything). As such we need to pass in
the hosts private ip address.
Once nomad 1.3.0 is out, this will actually be much easier since
variable interpolation will work for network.dns