Poll for consul to be ready
When using vault-generated certs, consul takes a bit of time to be
ready for services. If its not ready, nomad gets an odd constraint
error where it can't seem to read the consul version properly. The
right way to deal with this is to poll until consul is ready. We're
using the number of services as a quick n dirty way to check for
readiness. There should be a better way, so this will probably be
revisited in the future