Set scylla in single-node mode and add health check
By default scylla expects to run on overprovisioned nodes and will use
up as much memory and cpu as is available. That doesn't really make
sense for dev, so we're setting it in single-node mode via the smp flag
and pinned to 1 cpu core.
The health check is a bit uneven at the moment, ideally we should be
grepping for 'UN' in the result. It also seems to take a while for
scylla to come up, hence the 4 minute readiness check