Conditionally remove workaround for Cloudsmith promotion bug
To determine if Cloudsmith has addressed the underlying promotion bug,
we'll remove our workaround from the first time we run this "build and
upload images" job. This is not perfect, as we can't really know
whether the fix works from within this job, but it preserves the
workaround logic for now.

Tip for reviewers: view this PR with whitespace changes suppressed.

Fixes https://github.com/grapl-security/issue-tracker/issues/932

Signed-off-by: Christopher Maier <chris@graplsecurity.com>