Lower the sleep amount
I spent some time trying to find the right thing to poll on. I was
successful in polling for vault to be ready, but unfortunately the
main issue is that there's a race condition where Nomad can't get the
consul version. Nomad is able to actually assign the allocs but nomad
up exits due to the previous error even though Nomad can recover. Oh
well.