Use Python 3.8, managed by Pyenv
In an effort to increase reproducibility and consistency, we'll now
start managing our Python version through Pyenv. Pants is configured
configured to prefer Pyenv-managed Python versions, looking to the
version in the `.python-version` file, then to all Pyenv-managed
versions installed locally, and then finally to whatever Python
interpreter is found on `$PATH`.

Various lockfiles are updated to account for the new interpreter we
are using.

If you try running anything locally and Pants complains about not
being able to find a suitable Python interpreter, just run `pyenv
install` to build and use the version specified in `.python-version`.

(There is nothing particularly magical about Python 3.8 vs 3.7 or even
3.9 or 3.10, apart from the fact that it has some beneficial features
that we are using in other projects now.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>