Set up Vault and use Vault-generated certs for Consul (#1706)
This sets up Vault in dev and creates certs for Consul to better match what we're expecting to have in production. The main complexity of this change is around starting up the hashicorp stack in the right order and determining when the service is ready. There's also some minor improvements to the setup/teardown logic such as clearing the logs so that engineers are reading logs only for the current run's changes and not the logs from a previous build/run.