Add common shell library for checking script dependencies
We were checking in our `start_detach.sh` script to see if `nomad` and
`consul` are installed, but were missing a check to see if `vault` is
installed. Additionally, it can be a bit difficult to see in the
logging output if you're actually missing one of these binaries.

Here, we add a dedicated library file for dealing with such binary
checks, which can be reused across multiple scripts. It leverages our
colorized logging functions, as well as :sparkles: emoji :sparkles:,
making it very easy to find the cause of a failure (at least, as far
as required binaries are concerned).

Use of the `log_and_run` shell function is also added to
`start_detach.sh` to provide additional context, by logging the
command that is being run, and then running it.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>