Check if Nomad var is a pulumi.Output[str]
This fixes an issue where Nomad variables that were not string types
were being converted to strings. This is sort of ok for ints, but for
complex objects it causes problems. As such we only want to replace
pulumi.Output[str] with the fake pulumi strings. This was actually an
outstanding TODO previously, but we've now been incentivized to fix
it, even if it means using undocumented python internals :scream: