Internal 719 get generator rpc (#1784)
This PR makes some protocol buffer updates to simplify the plugin-registry's data model. We've consolidated the concept of PluginMetadata and removed the ability to download the plugin binary via RPC because it'll never be used.

This PR also implements the GetGeneratorsForEventSource RPC method, which necessitated a database schema migration to add event_source_id to the plugins table and associated updates to the data access layer.