Create library for to get hcl2 types via reflection.
This library will be used to fix type mismatches during
`pulumi preview`. This is necessary as during pulumi
preview, the pulumi object never awaits and resolves into
types. This results in us passing Pulumi objects into
Nomad for input variables that are expecting hcl2 types.

This can be seen when a PR creates a resource and then tries
to use a resource attribute with Nomad or alternatively
when a project needs to be built from scratch

 Conflicts:
	pulumi/infra/nomad_job.py