some improvements to the topic scanner
1. Now it panics instead of doing any error handling. Previously we were returning
errors but that was obscuring things like stack traces. Now we just panic, and
propagate that panic to the test via a JoinHandle.

2. Now it returns a JoinHandle<Vec<Envelope<T>>> containing all the Envelope<T>'s
matching the predicate instead of just the first match. This allows us to make
more interesting assertions about the results.

3. It's now a better Kafka citizen. Each test consumer gets its own unique
consumer group, and it commits its offsets.