Clarify workarounds in image build and upload logic
The logic for determining which images are actually new was behind a
retry gate (it would only run if the step was being retried in
Buildkite due to a failure). This unnecessarily coupled it to our
workaround for the fact that `docker buildx bake --push` seems to
sometimes cause issues for Cloudsmith, at least when many images are
being pushed at a time.

Our "is it new?" logic actually should **always** run. The bug that
caused us to add it in the first place appears to have been fixed, but
beyond that, the workaround is actually **better** for us, because it
means that we'll only end up updating running services when they
actually need to change, rather than every single time somebody merges
new code. The proper time to remove this logic is when we can a priori
determine which Rust-based container images actually need to be
built. Until that time, we need to keep it around.

Note: it's best to ignore whitespace differences when reviewing this
change. I'm unsure if the minute-long sleep is still required, but in
the interest of changing as little as possible, I opted to leave it in
for the time being.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>