Set up otel collector to send traces to Lightstep
This removes Jaeger in favor of Lightstep a SAAS. This means that when
we have cluster-wide problems we should still have telemetry to help
in debugging issues.
We are setting up otel-collector in order to take in traces in Zipkin
or Jaeger format, translate it to OTEL format and then forward it to
Lightstep. Ideally we would just use Vector for this, but Vector's
tracing support is in beta, and they don't support otel format :(
Since Lightstep has a secret access token, we're currently using
Pulumi to manage the secret since Vault still needs some additional
setup in order for Nomad to use it directly.