Use Consul DNS as dns server
With recent changes, we can now use a ip:port as the dns server. This
means we can get rid of dnsmasq or any other dns forwarding tool and
remove a source of complexity.
Consul DNS is already set to forward non Consul dns requests to whatever
is set as the server's dns. This means that .consul requests will be
replied to by the Consul agent and anything else will be replied to
by Amazon DNS.