Update pants run syntax after 2.13 change
As of pants 2.13 there's a change in how pants run works, which
requires us to now be more explicit about what's being run.

For reference:
https://buildkite.com/grapl/grapl-provision/builds/603#0183466b-f092-4e89-80c6-736920e24673

```
Currently, running a `pex_binary` by default will not include the source files in the PEX, and will instead put them in a temporary sandbox.

In Pants 2.14, the default will change to instead build the PEX like you had run the `package` goal, and then execute that PEX. This is more consistent and intuitive behavior.
```
This results in

```
TooManyTargetsException: The `run` goal only works with one valid target, but was given multiple valid targets:

  * etc/ci_scripts/dump_artifacts/__init__.py:dump_artifacts_lib
  * etc/ci_scripts/dump_artifacts/cli.py:dump_artifacts_lib
  * etc/ci_scripts/dump_artifacts/docker_artifacts.py:dump_artifacts_lib
  * etc/ci_scripts/dump_artifacts/nomad_artifacts.py:dump_artifacts_lib
  * etc/ci_scripts/dump_artifacts:dump_artifacts
```

Basically its looking at both the python_source targets and the pex_binary targets
and can't figure out which to use, so we need to be explicit