Nomad_endpoint must be a valid hostname
The nomad provider address is formatted like http://internalalb.foo.com:4646
This is an invalid hostname, because the http:// is for port 80. The
otel-collector blows up when unmarshalling the yaml config, because of
the invalid hostname. As such, we're stripping out the prepended http://
to make the alb a valid hostname.