Registry, not repository
TL;DR - it can be a subtle difference, but one that's important to
keep straight.

A Docker repository is a collection of related images, such as
`ubuntu`, `alpine`, or `hashicorp/vault`. A registry, on the other
hand, is a collection of repositories, such as Docker Hub.

This is complicated somewhat when we are also talking about artifact
storage services like Cloudsmith, which define "repository" as a
collection of artifacts, which can include "Docker repositories". That
is, I can have a Cloudsmith repository that contains the images
`foo:1.0`, `foo:1.1`, and `foo:1.2` (which would collectively define
the "foo" Docker repository), alongside `bar:1.0`, `bar:1.1`, and
`bar:1.2` (which collectively define the "bar" Docker
repository). Each image is an artifact to be managed, from
Cloudsmith's perspective.

Once it is time to actually pull an image from Cloudsmith, we have to
incorporate the Cloudsmith repository name into the proper Docker
_registry_ name, which will be something like
`docker.cloudsmith.io/${CLOUDSMITH_ORGANIZATION}/${CLOUDSMITH_REPOSITORY_NAME}`. To
this we would append an image name and tag.

(For simplicity, we generally refer to the Cloudsmith repository as
the combination of our organization name and the bare repository name,
e.g., `grapl/testing`. Cloudsmith itself appears to be a bit loose
with this distinction. Maybe we'll just think of it as a
"fully-qualified repository name".)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>