Add a `grapl` namespace to all container image names
We're looking at migrating to Artifactory, and this will make that a
bit easier. In particular, it allows us to have more control over
which images can be stored in a given repository, which brings both
security and performance benefits.

See, e.g., https://jfrog.com/blog/include-and-exclude-patterns/ for
further background on this.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>