Remove `protobuf` Python dependency pin
The problem here was that we were pinning to a specific `protoc`
version via Pants. This version was relatively old (the default
version that Pants comes with in version 2.13 is 3.20.1) and was
incompatible with the more recent versions of the Python `protobuf`
library.

By removing the Pants constraint, we can also remove the Python
constraint.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>