Set up head-based sampling
This sets up head-based sampling at the otel-collector layer. This
allows us to set a percentage of traces to collect. Currently we're
defaulting to 100% but we can decrease that as necessary.
Long-term, we should split the otel-collectors into agents and a
dedicated processing cluster. When we do that we should switch to
tail-based sampling so that we can create rules around what to sample