Fix s3 url format so that its explicitly s3
The hashicorp go-getter lib that Nomad uses internally doesn't
recognize the http:// uri as a s3 uri for some reason. As such, I'm
updating the format to be explicitly a s3 uri. This fixes the plugin
task when run in AWS
I've removed the comments about the slack thread, since that thread
has been deleted (past the 90 day threshold or whatever it is)