Use `jq` to extract the version of `protoc` that Pants uses
Rather than using `grep` on TOML to extract which version of `protoc`
is being used (which works, but can be fragile), we will instead parse
a gigantic JSON dump of Pants' state and extract the version using
`jq`.

This is more robust because it takes into account the layering that
Pants uses (i.e., CLI options override `pants.toml` configuration,
which in turn overrides hardcoded defaults). This means we don't even
have to pin a version in `pants.toml` and the script will still work!

For now, however, I *have* opted to keep the pin in `pants.toml`,
purely for visibility and explicitness.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>