Set up file exporting for otel
This fixes an issue where the file exporter couldn't create a file to
write to. This is because the official otel-collector container is
built off of the docker "scratch" container, which doesn't have a
writable /tmp directory. As such, we're using the distroless container
as a base, since that has a writable /tmp directory.