Allow playwright to be used by the grapl user
Playwright requires root access with the expectation that the user has
sudo permissions. As such the browsers end up being installed to the
user's home directory. As such, we're moving the files and chowning
them to the right owner.
Notably the other approach would be to give the grapl user sudo or su
permissions, however since the grapl user only exists to prevent
code being run as root, it doesn't quite make sense.